Refactor: ChatPage element를 App으로 끌어올림 + 종속 함수 재배치
