+++ src/hooks/Request/useSendRequestContext.ts
@@ -1,0 +1,6 @@
+import { useContext } from "react";
+import { SendRequestContext } from "./SendRequestContext";
+
+export default function useSendRequestContext() {
+	return useContext(SendRequestContext);
+}
